Cleanup and reuse coingecko code for ethereum generation
