--- packages/asset-service/src/generateAssetData/ethTokens/index.ts
@@ -1,1 +1,0 @@
-export * from './ethereum'
