+++ packages/asset-service/src/generateAssetData/ethereum/icons/foxy-icon.png
