+++ packages/asset-service/src/generateAssetData/ethereum/overrides.ts
@@ -1,5 +1,7 @@
 import { Asset } from '@shapeshiftoss/types'
 
+import { ethereum } from '../baseAssets'
+
 export const overrideTokens: Array<Asset> = [
   // example overriding FOX token with custom values instead of goingecko
   {
@@ -10,8 +12,8 @@ export const overrideTokens: Array<Asset> = [
     color: '#FFFFFF',
     icon: 'https://assets.coincap.io/assets/icons/fox@2x.png',
     symbol: 'FOX',
-    explorer: 'https://etherscan.io',
-    explorerAddressLink: 'https://etherscan.io/address/',
-    explorerTxLink: 'https://etherscan.io/tx/'
+    explorer: ethereum.explorer,
+    explorerAddressLink: ethereum.explorerAddressLink,
+    explorerTxLink: ethereum.explorerTxLink
   }
 ]
