+++ packages/asset-service/src/generateAssetData/ethereum/yearnVaults.ts
@@ -1,9 +1,21 @@
+import { JsonRpcProvider } from '@ethersproject/providers'
 import { ethChainId as chainId, toAssetId } from '@shapeshiftoss/caip'
 import { Asset } from '@shapeshiftoss/types'
 import { Token, Vault } from '@yfi/sdk'
+import { Yearn } from '@yfi/sdk'
 import toLower from 'lodash/toLower'
 
-import { yearnSdk } from './yearnSdk'
+import { ethereum } from '../baseAssets'
+
+const network = 1 // 1 for mainnet
+const provider = new JsonRpcProvider(process.env.REACT_APP_ETHEREUM_NODE_URL)
+export const yearnSdk = new Yearn(network, { provider })
+
+const explorerData = {
+  explorer: ethereum.explorer,
+  explorerAddressLink: ethereum.explorerAddressLink,
+  explorerTxLink: ethereum.explorerTxLink
+}
 
 export const getYearnVaults = async (): Promise<Asset[]> => {
   const vaults: Vault[] = await yearnSdk.vaults.get()
@@ -22,9 +34,7 @@ export const getYearnVaults = async (): Promise<Asset[]> => {
         assetNamespace: 'erc20',
         assetReference: vault.address
       }),
-      explorer: 'https://etherscan.io',
-      explorerAddressLink: 'https://etherscan.io/address/',
-      explorerTxLink: 'https://etherscan.io/tx/'
+      ...explorerData
     }
   })
 }
@@ -33,9 +43,7 @@ export const getIronBankTokens = async (): Promise<Asset[]> => {
   const ironBankTokens: Token[] = await yearnSdk.ironBank.tokens()
   return ironBankTokens.map((token: Token) => {
     return {
-      explorer: 'https://etherscan.io',
-      explorerAddressLink: 'https://etherscan.io/address/',
-      explorerTxLink: 'https://etherscan.io/tx/',
+      ...explorerData,
       color: '#276BDB', // yearn.finance blue
       icon: token.icon ?? '',
       name: token.name,
@@ -55,9 +63,7 @@ export const getZapperTokens = async (): Promise<Asset[]> => {
   const zapperTokens: Token[] = await yearnSdk.tokens.supported()
   return zapperTokens.map((token: Token) => {
     return {
-      explorer: 'https://etherscan.io',
-      explorerAddressLink: 'https://etherscan.io/address/',
-      explorerTxLink: 'https://etherscan.io/tx/',
+      ...explorerData,
       color: '#7057F5', // zapper protocol purple
       icon: token.icon ?? '',
       name: token.name,
@@ -77,9 +83,7 @@ export const getUnderlyingVaultTokens = async (): Promise<Asset[]> => {
   const underlyingTokens: Token[] = await yearnSdk.vaults.tokens()
   return underlyingTokens.map((token: Token) => {
     return {
-      explorer: 'https://etherscan.io',
-      explorerAddressLink: 'https://etherscan.io/address/',
-      explorerTxLink: 'https://etherscan.io/tx/',
+      ...explorerData,
       color: '#FFFFFF',
       icon: token.icon ?? '',
       name: token.name,
